requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
openai==1.3.7
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
sqlalchemy==2.0.23
python-dotenv==1.0.0

